+++ app/controllers/api/analytics_controller.rb
@@ -5,6 +5,8 @@ class API::AnalyticsController < API::ApiController
   before_action :authenticate_user!
 
   def data
+    authorize :analytics
+
     render json: HealthService.row_stats
   end
 end