+++ app/policies/analytics_policy.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Check the access policies for API::AnalyticsController
+class AnalyticsPolicy < ApplicationPolicy
+  def data?
+    user.admin?
+  end
+end